// web/src/gen/model/pinTypeMetadataDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PinTypeMetadataDto {
  code: string;
  name: string;
  category: string;
  color: string;
  defaultControl: string;
  /** @nullable */
  supportedEntityTargets?: unknown[] | null;
}